requirements.txt
click>=8.0
requests>=2.28
pyyaml>=6.0
